Tests/LibWeb/Text/expected/css-has-invalidation/has-featureless-and-concrete-child-list-coalescing.txt

[mixed featureless and class feature insertions] rgb(255, 0, 0)
[mixed featureless and id feature insertions] rgb(255, 0, 0)
[mixed featureless and attribute feature insertions] rgb(255, 0, 0)
[mixed featureless and tag feature insertions] rgb(255, 0, 0)
